Fix normalisation of the power spectrum.
Now we are normalising by mass rather than density, the power
spectrum divisor in powerspectrum.c needs to be the total mass in the
field (or in both fields if we are doing a cross-spectrum). This means
that read_fieldize for Gadget also needs to store something looking like
a mass, although for now I have set it equal to unity.